# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Ledgerline API
  version: 1.0.0
  description: >-
    API for document ingestion, retrieval, and agentic querying in the Ledgerline
    PageIndex pipeline.
servers:
  - url: http://localhost:8083
    description: Development (API Gateway)
  - url: https://api.yourdomain.com
    description: Production
...
paths:
...
  /documents/upload:
    post:
      tags: [Documents]
      summary: Upload a PDF document
      description: >-
        This endpoint is served by the ingestion service during local development.
      operationId: uploadDocument
      servers:
        - url: http://localhost:8080
          description: Development (Ingestion Service)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
  /documents:
    get:
      tags: [Documents]
      summary: List processed documents
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Documents list
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total number of documents
            X-Page-Count:
              schema:
                type: integer
              description: Total pages for current pagination strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsListResponse'
...
  /documents/{doc_id}/tree:
    get:
      tags: [Documents]
      summary: Get document tree structure
      operationId: getDocumentTree
      parameters:
        - $ref: '#/components/parameters/DocId'
      responses:
        '200':
          description: Tree retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTreeResponse'
        '404':
          $ref: '#/components/responses/DocumentNotFound'
...
components:
  parameters:
    DocId:
      name: doc_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Document UUID
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: -created_at
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: string
          example: error
        code:
          type: string
      required: [error, status, code]
    UploadResponse:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
        status:
          type: string
          example: uploaded
        message:
          type: string
          example: Document uploaded successfully
      required: [doc_id, status, message]
    Document:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
        filename:
          type: string
        uploaded_at:
          type: string
          format: date-time
        pages:
          type: integer
          minimum: 1
        status:
          type: string
          example: completed
        tree_generated_at:
          type: string
          format: date-time
      required: [doc_id, filename, uploaded_at, pages, status]
    DocumentsListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        count:
          type: integer
      required: [documents, count]
    TreeNode:
      type: object
      properties:
        title:
          type: string
        node_id:
          type: string
        start_index:
          type: integer
        end_index:
          type: integer
        summary:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/TreeNode'
    DocumentTreeResponse:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
        tree:
          $ref: '#/components/schemas/TreeNode'
      required: [doc_id, tree]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Missing required field
            status: error
            code: INVALID_REQUEST
    PayloadTooLarge:
      description: Uploaded file exceeds size limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Uploaded file exceeds size limit
            status: error
            code: FILE_TOO_LARGE
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unexpected server error
            status: error
            code: INTERNAL_ERROR
    DocumentNotFound:
      description: Document not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Document not found
            status: error
            code: DOCUMENT_NOT_FOUND
...